move QuantityFormatter Guardrails extension to LoopKitUI
Also move `CorrectionRangeOverrides.Preset.descriptiveText` extension to LoopKitUI
